refactor: Split immutableapp.js into multiple files
Large commit with code moved to actions/ and reducers/ folder. App.js
and src/index.js have updated imports. Some type signatures for
functions in actions/index.js were refactored prior to moving code.
